Extra code removed and updated readme file
